Making some modifications required to get triSYCL compiling with the SYCL component of SLAMBench. A little bit of a hackjob at the moment to get it to compile that needs a lot of tidying up.
Still unsure how best to integrate triSYCL in this case, trying to use another set of scripts to support the many options that triSYCL has when compiling and the fact that it can be used with and without a device compiler.

triSYCL also doesn't play nice with some of the older versions of ComputeCPP that use older/discarded SYCL specification components, need to decide to either add more macros to support backwards compatibility or to just keep a release that works with the older versions of ComputeCPP and only support newer ComputeCPP versions and triSYCL.

Currently it should compile if using the slambench_modifications branch of my triSYCL fork and the triSYCL-wip branches of dagr and syclcc.