Should work with triSYCL OpenMP now
Addition of debug functions and a temporary that involves making 
temporary copies of variables to pass to kernels while I work out the 
reason it doesn't work with triSYCL and apply a fix.

This won't technically work with triSYCL OpenMP until my changes to 
triSYCL are pushed and merged that involve changing the triSYCL maths 
header.

The main changes required were to triSYCL the vec classes were not sized 
appropriately due to an alignment issue causing reinterpret_cast's to 
fail, this is fixed and merged into triSYCL.

The other problems were missing and some slightly incorrectly 
implemented math functions for vec's, I have fixed these and they'll be 
pushed soon.

Another issue is with the script I made for triSYCL, I used -DOPENMP 
instead of -fopenmp which caused an issue with threading and barriers. 
This will be pushed to the syclcc scripts repository.

The final issue is something I hope to fix in the near future, triSYCL 
doesn't copy variables that are a member of kfusion across to the 
kernel's invoked by kfusion unless they're copied to a temporary 
variable and then the temporary is passed across. But this only seems to 
happen sometimes...

Future Steps for this branch are:

* Fix it so that it can be compiled for both ComputeCpp and triSYCL, as 
at the moment this won't compile with ComputeCpp as I cut some corners.
* Fix the triSYCL related issue so I can remove the work around