Modification of reduceKernel and Minor CMakeList modificaitons
Now use a plain SYCL parallel_workgroup for the reduce kernel rather 
than a DAGR parallel_for, this is to avoid explicit use of barriers 
which triSYCL's OpenMP implementation doesn't handle well (very slow). 
Doesn't seem to affect ComputeCpp performance.

Minor modifications to improve CMakeLists.txt. 

The first is adding -fopenmp which was missing and causing the -cpp 
implementation to run a lot slower than it should, at least on my 
machine at Xilinx. This appears to be missing in the actual SLAMBench 
repository as well, so maybe worth opening up an issue or merge request 
related to it.

Second is to change add_definitions -> target_compile_definitions, which 
means the definitions will only be present when compiling for the 
specific target. This means that depending on the syclcc script in use 
more SLAMBench modules are compileable...rather than having to swap 
between scripts. 

Full integration would be ideal, perhaps it's possible to swap to using 
the scripts when compiling for the sycl module...?